feat: variant version will be the default version + official variant
The official variant will pin the official binaries from
github:zen-browser/desktop, that's the option for people that doesn't
trust the binaries of this repository.

In other hand, the "twilight" package is pinned with artifacts that we
store in this repository. The reason why this is default now is because
I decided to follow the philosophy that is available in the official
NixOS page(https://nixos.org/): reproducible, declarative and reliable.
I know this could be a little bit invasive for some people but I think
that they're just a minority and this is a problem that should be solved
now given the time that has been passed since twilight was implemented.