feat(hm-module,spaces): implement places.sqlite updating via home.file.* script
This fixed the problem of the systemd service being rerun on every
home-manager rebuild. Now the places.sqlite updating script just runs
when something needs to be updated on it. We use home.file.* to store
the script on the home directory and take advantage of the
home.file.*.onChange option to run it.